+++ src/test/java/step3/CarGameTest.java
@@ -1,6 +1,7 @@
 package step3;
 
 import org.junit.jupiter.api.Test;
+import step3.game.CarGame;
 
 import java.util.Random;
