+++ src/test/java/step3/CarTest.java
@@ -3,6 +3,8 @@ package step3;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import step3.entity.Car;
+import step3.entity.Cars;
 
 import static org.assertj.core.api.Assertions.*;
