+++ src/test/java/step3/integration/CarGameTest.java
@@ -1,15 +1,14 @@
 package step3.integration;
 
 import org.junit.jupiter.api.Test;
-import step3.CarGame;
+import step3.game.CarGame;
 
 public class CarGameTest {
 
 
     @Test
-    void name() {
-        CarGame carGame = new CarGame(3,5);
-
+    void 자동차_게임_통합테스트() {
+        CarGame carGame = new CarGame(10,20);
         carGame.start();
     }
 }
